+++ R/zzz.R
@@ -10,5 +10,6 @@
   options("logR.mail_args" = NULL)
   options("logR.wd" = NULL)
   options("logR.silent" = TRUE)
+  options("logR.insert.returning" = NULL) # user provided function(ins.tab, ins.col, ins.val), see psql example: logR:::insert.returning.psql
   
 }